Rename `{keylog,keylogfile}` to `{key_log,key_log_file}`.
Be consistent with other modules.

```sh
git diff HEAD^1:rustls/src/keylog.rs rustls/src/key_log.rs
git diff HEAD^1:rustls/src/keylogfile.rs rustls/src/key_log_file.rs
```
