+++ rustls/src/lib.rs
@@ -329,8 +329,8 @@ mod check;
 mod bs_debug;
 mod builder;
 mod key;
-mod keylog;
-mod keylogfile;
+mod key_log;
+mod key_log_file;
 mod kx;
 mod suites;
 mod ticketer;
@@ -359,8 +359,8 @@ pub use crate::conn::{
 };
 pub use crate::error::Error;
 pub use crate::key::{Certificate, PrivateKey};
-pub use crate::keylog::{KeyLog, NoKeyLog};
-pub use crate::keylogfile::KeyLogFile;
+pub use crate::key_log::{KeyLog, NoKeyLog};
+pub use crate::key_log_file::KeyLogFile;
 pub use crate::kx::{SupportedKxGroup, ALL_KX_GROUPS};
 pub use crate::msgs::enums::CipherSuite;
 pub use crate::msgs::enums::ProtocolVersion;
